Add BLE UUIDs and improve debug functionality; refactor main.h for better structure
